renderFilteredItems tests file
